+++ image.h
@@ -33,8 +33,8 @@ ImagePtr loadimg(const char* file);
 ImagePtr loadppm(const char* file);
 uint8_t saveppm(const char *file, ImagePtr img);
 
-int8_t getpixel(ImagePtr img, uint16_t x, uint16_t y, uint16_t *pixel);
-int8_t setpixel(ImagePtr img, uint16_t x, uint16_t y, uint16_t *pixel);
+int8_t getpixel(ImagePtr img, uint16_t x, uint16_t y, uint8_t *pixel);
+int8_t setpixel(ImagePtr img, uint16_t x, uint16_t y, uint8_t *pixel);
 void printimg(ImagePtr img);
 
 #endif
